+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # SikuliX Changelog
 
+## 1.0.0
+  
+- Changed all option attribute defaults to false
+- Placed guard on setup to run only once
+- Added Remote Server
+
 ## 0.1.0
 
 - Initial release using sikulix v1.1.0 nightly builds